eliminate dupe profile name function
